# Docs/openapi/roodi.swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roodi API",
    "version": "1.0.0",
    "summary": "Contrato HTTP oficial do backend do projeto Roodi",
    "description": "Este contrato consolida os endpoints e payloads do projeto com base nas documentacoes internas:\n- 01PROJETO, 02STACK, 03REGRAS, 04FLUXOS, 07STRUCTURE, 08PAGES, 09MODULOS\n- API-INFINITY-PAY para integracao de pagamentos\n\nPremissas de negocio refletidas neste contrato:\n1) Sem geolocalizacao em tempo real no app.\n2) Sem mapa offline no fluxo principal.\n3) Tracking por estados/eventos.\n4) Regra de preco admin_only.\n5) Distancia/tempo por matriz de bairros + fallback deterministico.\n"
  },
  "servers": [
    {
      "url": "https://api.roodi.app",
      "description": "Producao"
    },
    {
      "url": "http://localhost:3333",
      "description": "Desenvolvimento local"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/auth/password/forgot": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Solicitar OTP de recuperacao de senha",
        "operationId": "forgotPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Solicitacao aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/password/otp/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verificar OTP",
        "operationId": "verifyOtp",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Redefinir senha com OTP verificado",
        "operationId": "resetPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Senha redefinida"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Consultar perfil autenticado",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Perfil do usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Me"
        ],
        "summary": "Atualizar perfil autenticado",
        "operationId": "updateMe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/me/settings/notifications": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Consultar preferencias de notificacao",
        "operationId": "getMyNotificationSettings",
        "responses": {
          "200": {
            "description": "Preferencias atuais",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Me"
        ],
        "summary": "Atualizar preferencias de notificacao",
        "operationId": "updateMyNotificationSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferencias atualizadas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/support/tickets": {
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "Listar chamados do usuario autenticado",
        "operationId": "listSupportTickets",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de chamados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Support"
        ],
        "summary": "Abrir chamado",
        "operationId": "createSupportTicket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportTicketCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chamado criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicket"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/tickets/{ticketId}": {
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "Consultar chamado",
        "operationId": "getSupportTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/TicketIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhe do chamado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicket"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/v1/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Listar usuarios",
        "operationId": "listAdminUsers",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/users/{userId}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Alterar status de usuario",
        "operationId": "updateAdminUserStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/pricing/rules": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Consultar regras de precificacao",
        "operationId": "getPricingRules",
        "responses": {
          "200": {
            "description": "Regras vigentes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRules"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Atualizar regras de precificacao",
        "operationId": "updatePricingRules",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingRulesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Regras atualizadas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRules"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/support/tickets": {
      "get": {
        "tags": [
          "Admin",
          "Support"
        ],
        "summary": "Listar chamados para operacao de suporte",
        "operationId": "listAdminSupportTickets",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de chamados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/support/tickets/{ticketId}": {
      "patch": {
        "tags": [
          "Admin",
          "Support"
        ],
        "summary": "Atualizar chamado (status, responsavel, observacao)",
        "operationId": "updateAdminSupportTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/TicketIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSupportTicketUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chamado atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicket"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "TicketIdParam": {
        "name": "ticketId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "UserIdParam": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Requisicao invalida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Recurso nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "rider",
          "commerce",
          "admin"
        ]
      },
...
      "SupportTicketStatus": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "closed"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ForgotPasswordRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "OtpChallengeResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "challenge_id",
              "expires_at",
              "resend_in_seconds"
            ],
            "properties": {
              "challenge_id": {
                "type": "string",
                "format": "uuid"
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "resend_in_seconds": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        }
      },
      "VerifyOtpRequest": {
        "type": "object",
        "required": [
          "challenge_id",
          "otp"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "otp": {
            "type": "string",
            "minLength": 4,
            "maxLength": 8
          }
        }
      },
      "OtpVerifyResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "reset_token",
              "expires_at"
            ],
            "properties": {
              "reset_token": {
                "type": "string"
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "reset_token",
          "new_password"
        ],
        "properties": {
          "reset_token": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
...
      "UserProfileResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
...
      "UserProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "address_home": {
            "$ref": "#/components/schemas/Address"
          },
          "address_base": {
            "$ref": "#/components/schemas/Address"
          },
          "bank_account": {
            "$ref": "#/components/schemas/BankAccount"
          },
          "vehicle": {
            "$ref": "#/components/schemas/Vehicle"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "cep": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          }
        }
      },
...
      "NotificationSettings": {
        "type": "object",
        "required": [
          "delivery",
          "payment",
          "promotions",
          "app_updates",
          "security",
          "support"
        ],
        "properties": {
          "delivery": {
            "type": "boolean"
          },
          "payment": {
            "type": "boolean"
          },
          "promotions": {
            "type": "boolean"
          },
          "app_updates": {
            "type": "boolean"
          },
          "security": {
            "type": "boolean"
          },
          "support": {
            "type": "boolean"
          }
        }
      },
      "NotificationSettingsUpdateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NotificationSettings"
          }
        ]
      },
...
      "SupportTicketCreateRequest": {
        "type": "object",
        "required": [
          "subject",
          "description",
          "priority"
        ],
        "properties": {
          "subject": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "order_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SupportTicket": {
        "type": "object",
        "required": [
          "id",
          "subject",
          "description",
          "status",
          "priority",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SupportTicketStatus"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "assigned_to": {
            "$ref": "#/components/schemas/UserSummary"
          }
        }
      },
      "SupportTicketListResponse": {
        "type": "object",
        "required": [
          "success",
          "data",
          "pagination"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportTicket"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/ApiPagination"
          }
        }
      },
...
      "AdminUserListResponse": {
        "type": "object",
        "required": [
          "success",
          "data",
          "pagination"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserProfile"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/ApiPagination"
          }
        }
      },
      "AdminUserStatusUpdateRequest": {
        "type": "object",
        "required": [
          "status",
          "reason"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "PricingRules": {
        "type": "object",
        "required": [
          "urgency_addon_brl",
          "conditional_addons_brl",
          "distance_zones_brl",
          "minimum_charge_brl",
          "max_distance_km"
        ],
        "properties": {
          "urgency_addon_brl": {
            "type": "object",
            "required": [
              "padrao",
              "urgente",
              "agendado"
            ],
            "properties": {
              "padrao": {
                "type": "number",
                "format": "float"
              },
              "urgente": {
                "type": "number",
                "format": "float"
              },
              "agendado": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "conditional_addons_brl": {
            "type": "object",
            "required": [
              "sunday",
              "holiday",
              "rain",
              "peak"
            ],
            "properties": {
              "sunday": {
                "type": "number",
                "format": "float"
              },
              "holiday": {
                "type": "number",
                "format": "float"
              },
              "rain": {
                "type": "number",
                "format": "float"
              },
              "peak": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "distance_zones_brl": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PricingZoneRule"
            }
          },
          "minimum_charge_brl": {
            "type": "number",
            "format": "float"
          },
          "max_distance_km": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "PricingRulesUpdateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PricingRules"
          }
        ]
      },
...
      "AdminSupportTicketUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SupportTicketStatus"
          },
          "note": {
            "type": "string"
          },
          "assigned_to_user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}